requirements.txt
streamlit==1.28.0
google-generativeai==0.3.1
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
deep-translator==1.11.4
gtts==2.4.0
reportlab==4.0.5
textblob==0.17.1